[Bridge/PhpUnit] Add PHPUnit 6 support
