+++ src/Symfony/Bridge/PhpUnit/Legacy/SymfonyTestsListener.php
@@ -1,0 +1,51 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Bridge\PhpUnit\Legacy;
+
+use Symfony\Bridge\PhpUnit\SymfonyTestsListenerTrait;
+
+/**
+ * Collects and replays skipped tests.
+ *
+ * @author Nicolas Grekas <p@tchwork.com>
+ *
+ * @internal
+ */
+class SymfonyTestsListener extends \PHPUnit_Framework_BaseTestListener
+{
+    use SymfonyTestsListenerTrait;
+
+    public function startTestSuite(\PHPUnit_Framework_TestSuite $suite)
+    {
+        return $this->doStartTestSuite($suite);
+    }
+
+    public function addSkippedTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
+    {
+        return $this->doAddSkippedTest($test, $e, $time);
+    }
+
+    public function startTest(\PHPUnit_Framework_Test $test)
+    {
+        return $this->doStartTest($test);
+    }
+
+    public function addWarning(\PHPUnit_Framework_Test $test, \PHPUnit_Framework_Warning $e, $time)
+    {
+        return $this->doAddWarning($test, $e, $time);
+    }
+
+    public function endTest(\PHPUnit_Framework_Test $test, $time)
+    {
+        return $this->doEndTest($test, $time);
+    }
+}
