+++ src/Symfony/Bridge/PhpUnit/Tests/DeprecationErrorHandler/default.phpt
@@ -18,13 +18,18 @@ require_once __DIR__.'/../../bootstrap.php';
 
 @trigger_error('root deprecation', E_USER_DEPRECATED);
 
-class PHPUnit_Util_Test
+eval(<<<'EOPHP'
+namespace PHPUnit\Util;
+
+class Test
 {
     public static function getGroups()
     {
         return array();
     }
 }
+EOPHP
+);
 
 class FooTestCase
 {
