+++ src/Symfony/Bridge/PhpUnit/composer.json
@@ -25,7 +25,7 @@
         "ext-zip": "Zip support is required when using bin/simple-phpunit"
     },
     "conflict": {
-        "phpunit/phpunit": ">=6.0"
+        "phpunit/phpunit": "<4.8.35|<5.4.3,>=5.0"
     },
     "autoload": {
         "files": [ "bootstrap.php" ],
